checkerrr/requirements.txt
# Core dependencies
aiogram>=3.0.0
python-dotenv>=1.0.0

# HTTP client for parsing ratings
aiohttp>=3.9.0

# Redis (optional, for production caching)
# redis>=5.0.0  # Для async: redis[hiredis]
# Или для aioredis:
# aioredis>=2.0.0

# Для fuzzy поиска (встроено в Python через difflib, но можно улучшить)
# thefuzz>=0.20.0  # Опционально для улучшенного fuzzy matching

# Для production
# uvloop>=0.19.0  # Для ускорения asyncio на Linux
# aiofiles>=23.0.0  # Для асинхронной работы с файлами
